+++ .gitignore
@@ -1,4 +1,5 @@
 /target
+/tools/repair-pnpm-lockfile/target/
 /node_modules
 /admin-ui-assets/*
 /playwright-report
