+++ Cargo.toml
@@ -106,7 +106,6 @@ utoipa = { version = "5.5.0", features = [
 ] }
 utoipa-swagger-ui = { version = "9.0.2", features = ["axum", "vendored"] }
 uuid = { version = "1.23.1", features = ["v7"] }
-yaml-rust = "0.4.5"
 
 [dev-dependencies]
 tempfile = "3.27.0"
