+++ src/lib.rs
@@ -61,7 +61,6 @@ pub mod images;
 pub mod middleware;
 pub mod migration;
 pub mod oidc;
-pub mod pnpm_lockfile_repair;
 pub mod publish;
 pub mod site_export;
 pub mod telemetry;
