+++ tools/repair-pnpm-lockfile/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "repair-pnpm-lockfile"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+clap = { version = "4.6.1", features = ["derive"] }
+serde_json = "1.0.150"
+thiserror = "2.0.18"
+yaml-rust = "0.4.5"
+
+[dev-dependencies]
+tempfile = "3.27.0"
